Fix task 13

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -2,7 +2,7 @@
 """ Module for task 13 """
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from model_state import State
+from model_state import Base, State
 import sys
 
 
@@ -10,12 +10,12 @@ if __name__ == "__main__":
     engine = create_engine("mysql+mysqldb://{}:{}@localhost:3306/{}"
                            .format(sys.argv[1], sys.argv[2], sys.argv[3]),
                            pool_pre_ping=True)
-    engine.connect()
     Session = sessionmaker(bind=engine)
     session = Session()
 
-    for state in session.query(State).filter(State.name.like('%a')):
-        session.delete(state)
+    for state in session.query(State):
+        if "a" in state.name:
+            session.delete(state)
     session.commit()
 
     session.close()
